requirements.txt
attrs==21.4.0
beancount==2.3.4
beautifulsoup4==4.10.0
bottle==0.12.19
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
docopt==0.6.2
et-xmlfile==1.1.0
feedparser==6.0.8
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
lxml==4.7.1
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
ply==3.11
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-magic==0.4.24
requests==2.26.0
rsa==4.8
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tika==1.24
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
yarg==0.1.9
